refactor(analysis): extract near-dup pair scoring
